Rename openmaterial to openmaterial3d (#12)
