+++ .gitignore
@@ -20,3 +20,4 @@ coverage.xml
 .coverage*
 .python-version
 generated_checker_bundle_doc.md
+xom_bundle_report.xqar
