+++ checker_bundle_doc.md
@@ -10,20 +10,20 @@
 
 ## Checkers
 
-### check_asam_openmaterial_valid_json_document
+### check_asam_xom_valid_json_document
 
 * Description: The given file to check must be a valid JSON document.
 * Addressed rules:
-  * asam.net:openmaterial:1.0.0:json.valid_json_document
+  * asam.net:xom:1.0.0:json.valid_json_document
 
-### check_asam_openmaterial_version_is_defined
+### check_asam_xom_version_is_defined
 
 * Description: The metadata of the file must contain an openMaterialVersion field.
 * Addressed rules:
-  * asam.net:openmaterial:1.0.0:json.version_is_defined
+  * asam.net:xom:1.0.0:json.version_is_defined
 
-### check_asam_openmaterial_valid_schema
+### check_asam_xom_valid_schema
 
 * Description: Input JSON file must be valid according to the corresponding schema.
 * Addressed rules:
-  * asam.net:openmaterial:1.0.0:json.valid_schema
+  * asam.net:xom:1.0.0:json.valid_schema
