+++ qc_openmaterial3d/__main__.py
@@ -4,6 +4,6 @@
 # Public License, v. 2.0. If a copy of the MPL was not distributed
 # with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from qc_openmaterial import main
+from qc_openmaterial3d import main
 
 main.main()
