+++ qc_openmaterial3d/schemas/1.0.0/asset_schema.json
