+++ qc_openmaterial3d/schemas/1.0.0/mapping_schema.json
