+++ qc_openmaterial3d/schemas/1.0.0/material_reflectance_schema.json
